backend/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
shortuuid==1.0.13
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
